Initial Commit: Weather forecast app
Initial commit includes this functionality:
A weather-card custom element that extends HTMLElement. It displays the current weather in the user's location which it gets with navigator.geolocation API. It then uses OpenWeatherMAP API to get the weather data 5 days in advance. It has added functionality to search for a custom location, it can display 5 of those which the user can select and see the weather forecast for those locations.
It also has the functionality to ask the user for an openweathermap API key that will be used later to display the info.